DANGEROUS! PROBLEMATIC! Make compile under 5.11
Well, it compiles...

But some serious changes has been made in the way vmap pointer is
exchanged between DRM parts. May be it has no effect here, but some
functions have to be revamped, or it may not work at all.

Just take a look at the changes in fl2000_gem_prime_vmap. WOW! Returning
a pointer for an integer is not that clever.

But well, it compiles...